refactor: separate FileUploadQuestion component and integrate it into FormQuestionRenderer
