+++ src/features/form/components/FormQuestionRenderer.tsx
@@ -1,122 +1,10 @@
-import * as formApi from "@/features/form/services/api";
 import { AccountButton, Checkbox, DateInput, DetailedCheckbox, DragToOrder, Input, Radio, ScaleInput, Select, TextArea } from "@/shared/components";
 import type { FormsQuestionResponse } from "@nycu-sdc/core-system-sdk";
-import { Chrome, Github, RefreshCw, Upload, X } from "lucide-react";
-import { useEffect, useRef, useState } from "react";
+import { Chrome, Github } from "lucide-react";
+import { useEffect } from "react";
+import { FileUploadQuestion } from "./FileUploadQuestion";
 import styles from "./FormDetailPage.module.css";
-
-const isValidUrl = (url: string): boolean => {
-	try {
-		const parsed = new URL(url);
-		return parsed.protocol === "http:" || parsed.protocol === "https:";
-	} catch {
-		return false;
-	}
-};
-
-interface FileItem {
-	id: string;
-	file: File;
-	status: "uploading" | "success" | "error";
-	error?: string;
-}
-
-interface FileUploadQuestionProps {
-	responseId?: string;
-	questionId: string;
-	maxFileAmount: number;
-	allowedFileTypes: string;
-	onFilesChange: (fileNames: string) => void;
-	disabled?: boolean;
-}
-
-const FileUploadQuestion = ({ responseId, questionId, maxFileAmount, allowedFileTypes, onFilesChange, disabled = false }: FileUploadQuestionProps) => {
-	const [items, setItems] = useState<FileItem[]>([]);
-	const inputRef = useRef<HTMLInputElement>(null);
-
-	const doUpload = async (item: FileItem) => {
-		if (!responseId) return;
-		setItems(prev => prev.map(i => (i.id === item.id ? { ...i, status: "uploading" as const, error: undefined } : i)));
-		try {
-			await formApi.uploadQuestionFiles(responseId, questionId, [item.file]);
-			setItems(prev => {
-				const next = prev.map(i => (i.id === item.id ? { ...i, status: "success" as const } : i));
-				onFilesChange(
-					next
-						.filter(i => i.status === "success")
-						.map(i => i.file.name)
-						.join(",")
-				);
-				return next;
-			});
-		} catch (err) {
-			setItems(prev => prev.map(i => (i.id === item.id ? { ...i, status: "error" as const, error: (err as Error).message } : i)));
-		}
-	};
-
-	const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-		if (!responseId || disabled) return;
-		const newFiles = Array.from(e.target.files || []);
-		const activeCount = items.filter(i => i.status !== "error").length;
-		const canAdd = Math.max(0, maxFileAmount - activeCount);
-		const toAdd: FileItem[] = newFiles.slice(0, canAdd).map(file => ({
-			id: crypto.randomUUID(),
-			file,
-			status: "uploading" as const
-		}));
-		if (inputRef.current) inputRef.current.value = "";
-		setItems(prev => [...prev, ...toAdd]);
-		toAdd.forEach(doUpload);
-	};
-
-	const removeItem = (id: string) => {
-		setItems(prev => {
-			const next = prev.filter(i => i.id !== id);
-			onFilesChange(
-				next
-					.filter(i => i.status === "success")
-					.map(i => i.file.name)
-					.join(",")
-			);
-			return next;
-		});
-	};
-
-	const activeCount = items.filter(i => i.status !== "error").length;
-	const canAddMore = !disabled && Boolean(responseId) && activeCount < maxFileAmount;
-
-	return (
-		<div className={styles.fileUploadList}>
-			{items.map(item => (
-				<div key={item.id} className={`${styles.fileItem} ${item.status === "error" ? styles.fileItemError : item.status === "uploading" ? styles.fileItemUploading : ""}`}>
-					<span className={styles.fileItemName}>{item.file.name}</span>
-					{item.status === "uploading" && <span className={styles.fileItemHint}>上傳中…</span>}
-					{item.status === "error" && <span className={styles.fileItemHint}>{item.error || "上傳失敗"}</span>}
-					<div className={styles.fileItemActions}>
-						{item.status === "error" && responseId && !disabled && (
-							<button type="button" className={styles.fileRetryBtn} onClick={() => doUpload(item)}>
-								<RefreshCw size={12} />
-								重新上傳
-							</button>
-						)}
-						{item.status !== "uploading" && (
-							<button type="button" className={styles.fileRemoveBtn} onClick={() => removeItem(item.id)} aria-label="移除" disabled={disabled}>
-								<X size={14} />
-							</button>
-						)}
-					</div>
-				</div>
-			))}
-			{canAddMore && (
-				<label className={styles.fileAddZone}>
-					<input ref={inputRef} type="file" multiple={maxFileAmount > 1} accept={allowedFileTypes} onChange={handleInputChange} className={styles.fileAddInput} />
-					<Upload size={14} />
-					<span>點擊上傳{maxFileAmount > 1 ? "（可多選）" : ""}</span>
-				</label>
-			)}
-		</div>
-	);
-};
+import { FormQuestionWrapper } from "./FormQuestionWrapper";
 
 type FormQuestionRendererProps = {
 	question: FormsQuestionResponse;
@@ -131,8 +19,6 @@ type FormQuestionRendererProps = {
 	onOtherTextChange: (questionId: string, value: string) => void;
 	onOauthConnect?: (question: FormsQuestionResponse) => void;
 };
-
-const getSelectedChoiceIds = (rawValue: string) => (rawValue ? rawValue.split(",").filter(Boolean) : []);
 
 export const FormQuestionRenderer = ({
 	question,
@@ -158,29 +44,30 @@ export const FormQuestionRenderer = ({
 		onAnswerChange(question.id, allowedIds[0]);
 	}, [question, sourceQuestion, sourceAnswerValue, value, onAnswerChange]);
 
+	const isValidUrl = (url: string): boolean => {
+		try {
+			const parsed = new URL(url);
+			return parsed.protocol === "http:" || parsed.protocol === "https:";
+		} catch {
+			return false;
+		}
+	};
+
+	const getSelectedChoiceIds = (rawValue: string) => (rawValue ? rawValue.split(",").filter(Boolean) : []);
+
 	switch (question.type) {
 		case "SHORT_TEXT":
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<Input id={question.id} placeholder="請輸入..." value={value} onChange={e => onAnswerChange(question.id, e.target.value)} required={question.required} />
-				</div>
+				</FormQuestionWrapper>
 			);
 
 		case "LONG_TEXT":
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<TextArea id={question.id} placeholder="請輸入..." value={value} onChange={e => onAnswerChange(question.id, e.target.value)} rows={6} required={question.required} />
-				</div>
+				</FormQuestionWrapper>
 			);
 
 		case "SINGLE_CHOICE": {
@@ -188,12 +75,7 @@ export const FormQuestionRenderer = ({
 			const otherChoice = choices.find(choice => choice.isOther);
 
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<Radio
 						options={choices.map(choice => ({ value: choice.id, label: choice.name }))}
 						value={value}
@@ -205,7 +87,7 @@ export const FormQuestionRenderer = ({
 						}}
 					/>
 					{otherChoice && value === otherChoice.id && <Input placeholder="請填寫其他" value={otherTextValue} onChange={e => onOtherTextChange(question.id, e.target.value)} />}
-				</div>
+				</FormQuestionWrapper>
 			);
 		}
 
@@ -213,14 +95,9 @@ export const FormQuestionRenderer = ({
 			const choices = question.choices ?? [];
 
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<Select options={choices.map(choice => ({ value: choice.id, label: choice.name }))} value={value || undefined} onValueChange={newValue => onAnswerChange(question.id, newValue)} />
-				</div>
+				</FormQuestionWrapper>
 			);
 		}
 
@@ -230,12 +107,7 @@ export const FormQuestionRenderer = ({
 			const selectedIds = getSelectedChoiceIds(value);
 
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<div className={styles.choiceList}>
 						{choices.map(choice => (
 							<Checkbox
@@ -254,18 +126,13 @@ export const FormQuestionRenderer = ({
 						))}
 					</div>
 					{otherChoice && selectedIds.includes(otherChoice.id) && <Input placeholder="請填寫其他" value={otherTextValue} onChange={e => onOtherTextChange(question.id, e.target.value)} />}
-				</div>
+				</FormQuestionWrapper>
 			);
 		}
 
 		case "DETAILED_MULTIPLE_CHOICE":
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<div className={styles.choiceList}>
 						{question.choices?.map(choice => (
 							<DetailedCheckbox
@@ -282,7 +149,7 @@ export const FormQuestionRenderer = ({
 							/>
 						))}
 					</div>
-				</div>
+				</FormQuestionWrapper>
 			);
 
 		case "DATE":
@@ -330,12 +197,7 @@ export const FormQuestionRenderer = ({
 			const displayIds = [...currentRankingIds, ...missingIds];
 
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					{shouldWaitSourceSelection ? (
 						<p className={styles.caption}>請先在「{sourceQuestion?.title || "來源題目"}」選擇至少一個項目，才可進行排序。</p>
 					) : (
@@ -352,18 +214,13 @@ export const FormQuestionRenderer = ({
 							}}
 						/>
 					)}
-				</div>
+				</FormQuestionWrapper>
 			);
 		}
 
 		case "UPLOAD_FILE":
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<FileUploadQuestion
 						responseId={responseId}
 						questionId={question.id}
@@ -376,17 +233,12 @@ export const FormQuestionRenderer = ({
 						最多 {question.uploadFile?.maxFileAmount || 1} 個檔案，每個檔案最大 {((question.uploadFile?.maxFileSizeLimit || 10485760) / 1024 / 1024).toFixed(0)} MB
 					</p>
 					{(!responseId || disableFileUpload) && <p className={styles.uploadHint}>預覽模式不會上傳檔案。</p>}
-				</div>
+				</FormQuestionWrapper>
 			);
 
 		case "OAUTH_CONNECT":
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<p className={styles.caption}>
 						綁定平台：
 						{question.oauthConnect === "GOOGLE" ? "Google" : "GitHub"}
@@ -403,17 +255,12 @@ export const FormQuestionRenderer = ({
 						</AccountButton>
 						<p className={styles.uploadHint}>{value ? `已綁定帳號：${value}` : "尚未綁定，點擊上方按鈕開始 OAuth 綁定流程。"}</p>
 					</div>
-				</div>
+				</FormQuestionWrapper>
 			);
 
 		case "HYPERLINK":
 			return (
-				<div key={question.id} className={styles.questionField}>
-					<label className={styles.questionLabel}>
-						{question.title}
-						{question.required && <span className={styles.requiredAsterisk}> *</span>}
-					</label>
-					{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+				<FormQuestionWrapper question={question}>
 					<Input
 						id={question.id}
 						placeholder="https://"
@@ -422,7 +269,7 @@ export const FormQuestionRenderer = ({
 						required={question.required}
 						error={value && !isValidUrl(value) ? "請輸入有效的網址（需以 http:// 或 https:// 開頭）" : ""}
 					/>
-				</div>
+				</FormQuestionWrapper>
 			);
 
 		default:
