+++ src/features/form/components/FormQuestionWrapper.tsx
@@ -1,0 +1,20 @@
+import type { FormsQuestionResponse } from "@nycu-sdc/core-system-sdk";
+import styles from "./FormQuestionWrapper.module.css";
+
+interface FormQuestionWrapperProps {
+	question: FormsQuestionResponse;
+	children: React.ReactNode;
+}
+
+export const FormQuestionWrapper = ({ question, children }: FormQuestionWrapperProps) => {
+	return (
+		<div key={question.id} className={styles.questionField}>
+			<label className={styles.questionLabel}>
+				{question.title}
+				{question.required && <span className={styles.requiredAsterisk}> *</span>}
+			</label>
+			{question.description && <div className={styles.questionDescription} dangerouslySetInnerHTML={{ __html: question.description }} />}
+			{children}
+		</div>
+	);
+};
